requirement.txt
pip list
```

You should see these key packages:
```
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-Limiter
PyMySQL
cryptography
langchain
langchain-google-genai
langchain-community
google-generativeai
python-dotenv
requests
chromadb
pytest


import os
os.environ['GEMINI_API_KEY'] = 'AIzaSyBIKxGCzg1rJE_OYxBLxb7IKDP6x-oTXfk'
from google import genai
client = genai.Client(api_key='AIzaSyBIKxGCzg1rJE_OYxBLxb7IKDP6x-oTXfk')
response = client.models.generate_content(
    model="gemini-2.0-flash",
    contents="Say hello in one word"
)
print(response.text)